refactor/rename integration tests to specify they're testing the `Queue` class; prep for future integration tests
