+++ integration_tests/common_queue_tests.py
@@ -16,17 +16,9 @@ import pytest  # type: ignore
 from MQClient import Queue
 from MQClient.backend_interface import Backend
 
+from .utils import DATA_LIST, _print_recv, _print_recv_multiple, _print_send
+
 logging.getLogger().setLevel(logging.DEBUG)
-
-
-# Note: don't put in duplicates
-DATA_LIST = [{'a': ['foo', 'bar', 3, 4]},
-             1,
-             '2',
-             [1, 2, 3, 4],
-             False,
-             None
-             ]
 
 
 @pytest.fixture  # type: ignore
@@ -37,27 +29,8 @@ def queue_name() -> str:
     return name
 
 
-def _print_recv(data: Any) -> None:
-    _print_data("RECV", data)
-
-
-def _print_recv_multiple(data: Any) -> None:
-    _print_data("RECV", data, is_list=True)
-
-
-def _print_send(data: Any) -> None:
-    _print_data("SEND", data)
-
-
-def _print_data(_type: str, data: Any, is_list: bool = False) -> None:
-    if (_type == "RECV") and is_list and isinstance(data, list):
-        print(f"{_type} - {len(data)} :: {data}")
-    else:
-        print(f"{_type} :: {data}")
-
-
-class PubSub:
-    """Integration test suite."""
+class PubSubQueue:
+    """Integration test suite for Queue objects."""
 
     backend = None  # type: Backend
