+++ integration_tests/test_apachepulsar.py
@@ -6,10 +6,10 @@ Verify basic functionality.
 # local imports
 from MQClient.backends import apachepulsar
 
-from .common_tests import PubSub, queue_name  # noqa: F401 # pylint: disable=W0611
+from .common_queue_tests import PubSubQueue, queue_name  # noqa: F401 # pylint: disable=W0611
 
 
-class TestPulsar(PubSub):
-    """Run PubSub integration tests with Pulsar backend."""
+class TestPulsarQueue(PubSubQueue):
+    """Run PubSubQueue integration tests with Pulsar backend."""
 
     backend = apachepulsar.Backend()
